Final version
Math and comment updates. Easily add people by just copying the row with the functions inside from someone else and paste the, and edit it to the new persons name.
Warning: there can ONLY be 1 unique name in the board. Their name is their UniqueID so if there is another person with the same name we will have to use part of last name